+++ src/components/layout/YcHeaderBox.vue
@@ -1,34 +1,20 @@
 <template>
-  <el-row type="flex" align="middle" :gutter="10">
-    <el-col :span="10">
-      <i :class="'el-icon-exp-bars yc-collapse-btn '  + (collapse ? ' is-active' : '')"
-         @click="collapseChange"></i>
-      <el-breadcrumb class="yc-title-breads" separator="/">
-        <el-breadcrumb-item :to="{name: 'Home'}">首页</el-breadcrumb-item>
-        <el-breadcrumb-item v-for="(title, index) in activeFullTitle" :key="index">
-          {{ title }}
-        </el-breadcrumb-item>
-      </el-breadcrumb>
-    </el-col>
-    <el-col :span="8" style="text-align: right;">
-      <el-link class="top-bar-link" icon="el-icon-s-help" :underline="false" target="_blank"
-               :href="$CONFIG.web.blog">个人博客</el-link>
-      <el-link class="top-bar-link" icon="el-icon-s-help" :underline="false" target="_blank"
-               :href="$CONFIG.web.github">GitHub</el-link>
-    </el-col>
-    <el-col :span="6" style="text-align: right;">
-      <yc-header-bar-box-right></yc-header-bar-box-right>
-    </el-col>
+  <el-row type="flex" :gutter="10" style="height: 60px;">
+    <yc-header-bar-box-left class="bar-box is-left"></yc-header-bar-box-left>
+    <yc-header-bar-box-mid class="bar-box is-mid"></yc-header-bar-box-mid>
+    <yc-header-bar-box-right class="bar-box is-right"></yc-header-bar-box-right>
   </el-row>
 </template>
 
 <script>
 import StoreUtils from '../../plugins/utils/StoreUtils'
 import YcHeaderBarBoxRight from './YcHeaderBarBoxRight'
+import YcHeaderBarBoxLeft from './YcHeaderBarBoxLeft'
+import YcHeaderBarBoxMid from './YcHeaderBarBoxMid'
 
 export default {
   name: 'YcHeaderBox',
-  components: { YcHeaderBarBoxRight },
+  components: { YcHeaderBarBoxMid, YcHeaderBarBoxLeft, YcHeaderBarBoxRight },
   computed: {
     activeFullTitle: function () {
       let fullTitle = []
@@ -55,32 +41,24 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-  .yc-collapse-btn {
-    line-height: 60px;
-    font-size: 32px;
-    margin-left: 10px;
-    cursor: pointer;
-    transform: rotate(0deg);
-    transition: 0.3s;
-    transform-origin: 50% 50%;
-
-    &.is-active {
-      transform: rotate(90deg);
+  .bar-box {
+    height: 60px;
+    align-items: center;
+    &.is-left {
+      display: flex;
     }
-  }
-
-  .yc-title-breads {
-    line-height: 60px;
-    position: absolute;
-    left: 60px;
-    display: inline-block;
-  }
-
-  .el-dropdown-link {
-    cursor: pointer;
-  }
-
-  .top-bar-link {
-    margin: 0 5px;
+    &.is-mid {
+      box-sizing: border-box;
+      display: flex;
+      padding: 0 50px;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+    }
+    &.is-right {
+      position: absolute;
+      right: 20px;
+      display: flex;
+    }
   }
 </style>
